refactor(g-base): merge AbstractContainer and ContainerUtil, close #180
